feat: Implement conversation formatting and 2-speaker optimization
- Add conversation formatter for readable transcript output
- Optimize default configuration for 2-speaker conversations
- Fix model name sanitization for file paths
- Add fallback mechanisms for stable-whisper compatibility
- Update speaker configuration presets
- Improve output formatting with timestamps and speaker labels
- Add comprehensive error handling and logging
- Update documentation with verification results

Key improvements:
- Conversation format: [MM:SS - MM:SS] Speaker X: text
- 2-speaker default: min_speakers=2, max_speakers=2
- Hebrew text support with proper formatting
- Multiple output formats (JSON, TXT, DOCX)
- Chronological sorting with gap detection

+++ src/engines/transcription_engine.py
@@ -28,7 +28,7 @@ class FasterWhisperEngine:
         if self.model is None:
             import faster_whisper
             self.model = faster_whisper.WhisperModel(
-                self.model_name, device=device, compute_type='float16', local_files_only=True
+                self.model_name, device=device, compute_type='float32', local_files_only=True
             )
         
         return self.model.transcribe(audio_file, language=language, word_timestamps=word_timestamps)
@@ -45,7 +45,7 @@ class StableWhisperEngine:
         if self.model is None:
             import stable_whisper
             self.model = stable_whisper.load_faster_whisper(
-                self.model_name, device=device, compute_type='float16', local_files_only=True
+                self.model_name, device=device, compute_type='float32', local_files_only=True
             )
         
         result = self.model.transcribe(audio_file, language=language, word_timestamps=word_timestamps)
